link accounts and courses with student.

+++ routes/accounts.js
@@ -1,23 +1,33 @@
 var express = require("express");
-var router = express.Router();
+var router = express.Router({mergeParams : true});
 var Account = require("../models/account");
+var Student = require("../models/student");
 var middleware = require("../middleware");
 var { isLoggedIn, isAdmin } = middleware;
 
 // INDEX Show all Accounts.
 router.get("/", (req, res) => {
+    // console.log("INDEX route : /" + req.params.acc_parent_type + "/" + req.params.acc_parent_id);
     Account.find({}, (err, accounts) =>{
         if (err) {
             console.log(err);
         } else {
-            res.render("accounts/index", { accounts : accounts , page: 'accounts'});
+            res.render("accounts/index", {
+                accounts : accounts,
+                page: 'accounts',
+                acc_parent_type : req.params.acc_parent_type,
+                acc_parent_id: req.params.acc_parent_id
+            });
         }
     });
 });
 
 // NEW render new page to create new account
 router.get("/new", isLoggedIn, (req, res) => {
-    res.render("accounts/new");
+    res.render("accounts/new", {
+        acc_parent_type : req.params.acc_parent_type,
+        acc_parent_id: req.params.acc_parent_id
+    });
 });
 
 // CREATE new account to be created and updated in DB:
@@ -26,8 +36,30 @@ router.post("/", isLoggedIn, (req, res)=>{
         if (err) {
             req.flash("error", "New account cannot be created error : " + err.message);
             console.log(err);
+            res.redirect("back");
         }
-        res.redirect("/accounts");
+        else {
+            // check whether request came from parent 
+            if ( (typeof req.params.acc_parent_type !== 'undefined') && (typeof req.params.acc_parent_id !== 'undefined') ) {
+                // add account to the student model also.
+                Student.findById(req.params.acc_parent_id, (err, student) => {
+                    if (err) {
+                        console.log(err);
+                        req.flash("error", "Student not found error : " + err.message);
+                        res.redirect("back");
+                    } else {
+                        student.account.id = account._id;
+                        student.account.name = account.name;
+                        student.save();
+                        res.redirect("/" + req.params.acc_parent_type + "/" + req.params.acc_parent_id);
+                    }
+                });
+            } else {
+                // simply display the accounts details
+                res.redirect("/accounts");        
+            }
+        }
+        
     });
 });
 
@@ -39,7 +71,11 @@ router.get("/:account_id", (req, res) => {
             req.flash("error", "Account not found " + err.message );
             res.redirect("back");
         } else {
-            res.render("accounts/show", {account : account});
+            res.render("accounts/show", {
+                account : account,
+                acc_parent_type : req.params.acc_parent_type,
+                acc_parent_id: req.params.acc_parent_id
+            });
         }
     });
 });
@@ -53,7 +89,11 @@ router.get("/:account_id/edit", isLoggedIn, (req, res) =>{
             res.redirect('back');
         }
         else {
-            res.render("accounts/edit", { account : account });
+            res.render("accounts/edit", {
+                account : account,
+                acc_parent_type : req.params.acc_parent_type,
+                acc_parent_id: req.params.acc_parent_id
+            });
         }
     });
 });
@@ -74,13 +114,39 @@ router.put("/:account_id", isLoggedIn, (req, res) =>{
 
 // DESTROY route : this will destroy the route.
 router.delete("/:account_id", isLoggedIn, (req, res) => {
-    Account.findByIdAndRemove(req.params.account_id, (err) => {
-        if (err) {
-            console.log(err);
-            req.flash("error", "Could not delete the account : " + err.message);
+    if ( (typeof req.params.acc_parent_type !== 'undefined') && (typeof req.params.acc_parent_id !== 'undefined') ) {
+        if(req.params.acc_parent_type == 'students') {
+            // Delete accounte for student.
+            Account.findByIdAndRemove(req.params.account_id, (err) => {
+                if (err) {
+                    console.log(err);
+                    req.flash("error", "Could not delete the account : " + err.message);
+                    res.redirect("back");
+                } else {
+                    Student.findById(req.params.acc_parent_id, (err, student) => {
+                        if (err) {
+                            console.log(err);
+                            req.flash("error", "Student not found error : " + err.message);
+                            res.redirect("back");
+                        } else {
+                            student.account = {};
+                            student.save();
+                            res.redirect("/" + req.params.acc_parent_type + "/" + req.params.acc_parent_id);
+                        }
+                    });
+                }
+            });
+        } else if(req.params.acc_parent_type == 'courses') {
+            req.flash("error", "Deleting account for course is not permitted.")
+            res.redirect("/courses");
+        } else {
+            req.flash("error", "Default deleting of account is not permitted");
+            res.redirect("back");
         }
-        res.redirect("/accounts");
-    });
+    } else {
+        req.flash("error", "To delete account use account holder's path");
+        res.redirect("back");
+    }
 });
 
 module.exports = router;
